+++ src/main/java/divinerpg/items/ranged/staffs/VetheanStaff.java
@@ -2,11 +2,9 @@ package divinerpg.items.ranged.staffs;
 
 import divinerpg.entities.projectile.magic.EntityBouncingProjectile;
 import divinerpg.items.ranged.ItemRangedWeapon;
-import divinerpg.registries.EntityRegistry;
-import divinerpg.registries.SoundRegistry;
+import divinerpg.registries.*;
 import divinerpg.util.LocalizeUtils;
-import net.minecraft.world.entity.EntityType;
-import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.*;
 import net.minecraft.world.entity.projectile.Projectile;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
@@ -14,8 +12,8 @@ import net.minecraft.world.level.Level;
 import java.util.function.Supplier;
 
 public class VetheanStaff extends ItemRangedWeapon {
-    float damage;
-    public VetheanStaff(int arcanaUse, float damage) {
+    int damage;
+    public VetheanStaff(int arcanaUse, int damage) {
         super(EntityRegistry.BOUNCING_PROJECTILE::value);
         arcanaConsumedUse = arcanaUse;
         this.damage = damage;
@@ -23,7 +21,7 @@ public class VetheanStaff extends ItemRangedWeapon {
         tooltips.add(LocalizeUtils.arcanaDam(damage));
         tooltips.add(LocalizeUtils.bouncingShots());
     }
-    protected VetheanStaff(Supplier<EntityType<? extends Projectile>> projectileType, int arcanaUse, float damage) {
+    protected VetheanStaff(Supplier<EntityType<? extends Projectile>> projectileType, int arcanaUse, int damage) {
         super(projectileType);
         arcanaConsumedUse = arcanaUse;
         this.damage = damage;
@@ -31,8 +29,7 @@ public class VetheanStaff extends ItemRangedWeapon {
         tooltips.add(LocalizeUtils.arcanaDam(damage));
         tooltips.add(LocalizeUtils.bouncingShots());
     }
-    @Override
-    protected EntityBouncingProjectile createProjectile(Level level, LivingEntity shooter, ItemStack weapon, ItemStack ammo, boolean isCrit) {
+    @Override protected EntityBouncingProjectile createProjectile(Level level, LivingEntity shooter, ItemStack weapon, ItemStack ammo, boolean isCrit) {
         EntityBouncingProjectile p = (EntityBouncingProjectile) super.createProjectile(level, shooter, weapon, ammo, isCrit);
         p.baseDamage = damage;
         return p;
